+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
+### Added Public Features
+### Added
+- [OMFErrorHandler] Improving ErrorManagement: 
+    - Take into account the ECriticality on the notification window
+    - does not display anything if ECriticality is SILENT
+
+### Fixed
+### Changed
 
 ## [1.0.25] - 2023-08-17